changed bday_log to contain strings matching current_date_lst format
changed the existing bday_log birthdays to be strings. if you want to compare the birthdays in bday_log to the date in current_date_lst then the format needs to be the same. an int and a str will never be equal and will never resolve to True in line 29.